+++ R/stat-alluvium.r
@@ -37,7 +37,7 @@ stat_alluvium <- function(mapping = NULL,
                           position = "identity",
                           decreasing = NA,
                           reverse = TRUE,
-                          relevel.strata = NULL,
+                          discern = FALSE,
                           aggregate.wts = FALSE,
                           lode.guidance = "zigzag",
                           lode.ordering = NULL,
@@ -56,7 +56,7 @@ stat_alluvium <- function(mapping = NULL,
     params = list(
       decreasing = decreasing,
       reverse = reverse,
-      relevel.strata = relevel.strata,
+      discern = discern,
       aggregate.wts = FALSE,
       lode.guidance = lode.guidance,
       lode.ordering = lode.ordering,
@@ -121,14 +121,14 @@ StatAlluvium <- ggproto(
     if (type == "alluvia") {
       axis_ind <- get_axes(names(data))
       data <- to_lodes(data = data, axes = axis_ind,
-                       relevel.strata = params$relevel.strata)
+                       discern = params$discern)
       # positioning requires numeric 'x'
       data <- data[with(data, order(x, stratum, alluvium)), , drop = FALSE]
       data$x <- contiguate(data$x)
     } else {
-      if (!is.null(params$relevel.strata)) {
+      if (!is.null(params$discern)) {
         warning("Data is already in lodes format, ",
-                "so 'relevel.strata' will be ignored.")
+                "so 'discern' will be ignored.")
       }
     }
     
@@ -137,7 +137,7 @@ StatAlluvium <- ggproto(
   
   compute_panel = function(data, scales,
                            decreasing = NA, reverse = TRUE,
-                           relevel.strata = NULL,
+                           discern = FALSE,
                            aggregate.wts = FALSE,
                            lode.guidance = "zigzag",
                            aes.bind = FALSE,