+++ R/stat-stratum.r
@@ -23,8 +23,8 @@
 #'   so that they match the order of the values in the legend.
 #'   Ignored if \code{decreasing} is not \code{NA}.
 #'   Defaults to \code{TRUE}.
-#' @param relevel.strata Passed to \code{\link{to_lodes}} if \code{data} is in
-#'   alluvia format.
+#' @param discern Passed to \code{\link{to_lodes}} if \code{data} is in alluvia
+#'   format.
 #' @param label.strata Logical; whether to assign the values of the axis 
 #'   variables to the strata. Defaults to FALSE, and requires that no
 #'   \code{label} aesthetic is assigned.
@@ -36,7 +36,7 @@ stat_stratum <- function(mapping = NULL,
                          position = "identity",
                          decreasing = NA,
                          reverse = TRUE,
-                         relevel.strata = NULL,
+                         discern = FALSE,
                          label.strata = FALSE,
                          show.legend = NA,
                          inherit.aes = TRUE,
@@ -53,7 +53,7 @@ stat_stratum <- function(mapping = NULL,
     params = list(
       decreasing = decreasing,
       reverse = reverse,
-      relevel.strata = relevel.strata,
+      discern = discern,
       label.strata = label.strata,
       na.rm = na.rm,
       ...
@@ -103,14 +103,14 @@ StatStratum <- ggproto(
     if (type == "alluvia") {
       axis_ind <- get_axes(names(data))
       data <- to_lodes(data = data, axes = axis_ind,
-                       relevel.strata = params$relevel.strata)
+                       discern = params$discern)
       # positioning requires numeric 'x'
       data <- data[with(data, order(x, stratum, alluvium)), , drop = FALSE]
       data$x <- contiguate(data$x)
     } else {
-      if (!is.null(params$relevel.strata)) {
+      if (!is.null(params$discern)) {
         warning("Data is already in lodes format, ",
-                "so 'relevel.strata' will be ignored.")
+                "so 'discern' will be ignored.")
       }
     }
     
@@ -124,7 +124,7 @@ StatStratum <- ggproto(
   
   compute_panel = function(self, data, scales,
                            decreasing = NA, reverse = TRUE,
-                           relevel.strata = NULL, label.strata = FALSE) {
+                           discern = FALSE, label.strata = FALSE) {
     
     # introduce label (if absent)
     if (label.strata) {