+++ NEWS.md
@@ -1,10 +1,10 @@
 # ggalluvial 0.5.0.9002
 
 - BC: Change `keep` parameter to `diffuse` (more mnemonic with `distill`).
+- `to_lodes()` now has a logical `discern` parameter that uses `make.unique()` to make stratum values that appear at different axes distinct. The `stat_*()` functions can pass the same parameter internally and print a warning if the data is already in lodes form.
 
 # ggalluvial 0.5.0.9001
 
-- `to_lodes()` now has a `relevel.strata` parameter to sort or customize the factor levels of the variable containing the aggregated strata. The `stat_*()` functions can pass the same parameter internally and print a warning if the data is already in lodes form.
 - `GeomFlow$draw_panel()` now begins by removing `complete.cases()`, corresponding to one-sided flows. (These are not removed in `StatFlow$compute_panel()`, which would exclude missing aesthetic values from legends.)
 - `to_alluvia()` now has a `distill` parameter to control the inclusion of any original variables that vary within values of `id` into the reformatted data, based on a distilling function that returns a single value from a vector.
 - `is_alluvial_alluvia` now prints a message rather than a warning when some combinations of strata are not linked by any alluvia.